backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
filelock==3.12.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.0
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rpds-py==0.9.2
SQLAlchemy==2.0.20
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
Werkzeug==2.3.7
yarl==1.9.2
